move SQLite UDF registration to a subscriber
